requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.10
streamlit==1.41.0
passlib==1.7.4
